requirements.txt
ipykernel
openai
python-dotenv
yfinance
pandas
matplotlib
seaborn
plotly
scikit-learn
scipy
statsmodels
colorama